Adding :app-framework module
This module aims to encapsulate framework components for the app target. This will be things like services, caches, navigation and implementations for things that we need to platform to provide like permissions etc. You could package these in the target :app module, however providing them as a different module will give you a compile time benefit on 2 fronts, incremental, as these things will be unlikely to change, and concurrent as they can be compiled along with the other modules. Generally this module will depend on all other modules, data (for services / cache), domain (for things like permissions although this should be avoided as much as possible), presentation (for navigation / permissions etc)