Adding ServiceFallbackCatFactsRepository
This CatFactsRepository implementation loads cached data falling back to the service if nothing is cached (null). This has the disadvantage of using stale data - although you could use a cache implementation that only returns data stored with in a time frame to overcome this. The positive side is that we reduce network bandwidth and have a faster implementation as we avoid network latency