Adding CoroutineJobManager
This manager allows you to launch coroutines under certain conditions, namely to not run the provided key if a job with the same key is currently in progress, or to cancel the job and replace with the provided job. This is particularly useful for handling actions triggered by the user, such as clicks, so can be used in ViewModels. For example approving an action, you would want to avoid double calls to do this so could simply ignore subsequent requests.

Don't forget to call cancel when finished - if using a view model hook this call up to onCleared to avoid leaking coroutines