Adding CoroutineRequestManager
This object aims to provide an API to execute Async tasks. If a client requests a task with the same name and return type they are returned the in progress task. This is useful for situation where you do not want to execute a task if one is already running and would rather get the result of the in progress task. A good example is a load call on a service, or a get action on a repository.

Unfortunately we need to do an unchecked cast to make this work, however we can guarantee the safety by using the types name as part of the key. I chose to use a mutex over the concurrent hash map as we read and add at different points. Without this I believe it could be possible for another coroutine to execute an add between a read and write - despite the map being concurrent