Adding LiveData Extension file with initial mutableLiveData function
This file exists to store top level and extension functions associated with LiveData. mutableLiveData creates a MutableLiveData instance and immediately assigns a default value. This is useful when implementing viewmodels as you can online assign the initial value - even at the assignment site, rather than declaring the object and the having to assign the value in the init method