Adding Navigator interface and LifecycleOwnerManager classes
In order to make components reusable across different targets and to aid with testing we want to abstract our navigation logic from our view model and even from our activities and fragments. To do this we would rout any events from the view model into a navigator interface. However for that navigator to actually navigate it would need a reference to either an activity or fragment. This is not available at injection time and even so would likely leak due to the way lifecycles work.

To get around this we have a base navigator interface that our view models can expose publicly, it contains LifecycleObserver property that the containing activity/fragment can bind as a listener eg

`this.lifecycle.addObserver(viewModel.navigator.lifecycleObserver)`

At this point we can get a reference to the LifecycleObserver. It still up to us to manage this and not leak the lifecycle. To this end we have AbstractActivityLifecycleOwnerManager. It registers a callback onStart and sets an internal instance of the lifecycleOwner. OnDestroy we null it and remove ourselves as an observer. We can then create subclasses that provide access to the concrete type. using generics and Class matching the generic type we can handle the cast internally in the abstract class - leaving the implementing class to simply specify the concrete type.

This is of course potentially dangerous, so we must be careful to provide the navigator with the correct AbstractLifecycleOwnerManager for the context it is being used in. However, because specific navigator implementations will be so closely tied to this context we cannot abstract it completely - this just helps reduce boiler plate and allow us to reuse view models in different contexts (we would provide different navigator implementations depending on whether the VM is being used in a fragment vs an activity for example)