Adding Coroutine Test Rule
This rule sets the main dispatcher for coroutine. This is useful when the code under test does not control the dispatcher it is using. The most obvious place this happens is in a view model that uses ViewModel Scope. In this scenario, to control the scope while testing we can apply this rule. We can then access, or set, the TestDispatcher and, if using TestDispatcher, use methods like advanceTime by. This is a must when testing state based code like view models