Adding CatFactSummariesPresentationModule
This module provides the CatFactSummariesViewModel. There are a few tricks to getting this done. Firstly we need to use our ViewModelKey to associate the viewModel interface with a ViewModel, we receive an instance of the concrete SimpleCatFactSummariesViewModel and return Viewmodel. Because we use @binds @intoMap and ViewModel key this is an abstract method - generated by dagger. Next we need to map The SimpleCatCatSummariesViewModel to the interface CatCatSummariesViewModel using @Binds. Finally we need a provider for the actual concrete instance. This is provided in a companion object dues to the fact in Dagger a abstract module cannot contain non abstract methods - we could us e a different class but because they are so tightly related it makes sense for them to live in the same file