Adding ApplicationActivityModule
This module aims to provide activity scopes that are one level below the Application / Singleton scope. IN this case we only need to provide for our CatFactSummariesActivity initially. Firstly we use @ContributesAndroidInjector to create a subcomponent for the activity, which receivers our presentation and navigation modules plus the ViewModel provider module which scopes the presentation layer to the activity